+++ tests/test_files/error_files/py_file_type_field_removal.py
@@ -1,0 +1,16 @@
+"""This is an example python file"""
+
+
+class ExampleClass:
+    def __init__(self, *args, **kwargs):
+        """First doc str
+
+        :type kwargs: dict
+        :param args: Args
+        :param kwargs: Kwargs but with a really long description that will need to be
+            rewrapped because it is really long and won't fit in the default 88
+            characters.
+        :returns: Returns
+        :type args: list
+
+        """