+++ CHANGES.rst
@@ -4,6 +4,26 @@ Change Log
 Unreleased
 ----------
 
+**Added**
+
+- Added more missing roles.
+- Added support for Python 3.11.
+
+**Changed**
+
+- Improved field sorting and formatting.
+- Improved handling of ``:param:`` and ``:type:`` fields.
+- Bumped minimum version of ``docutils`` to 0.19.*.
+- Bumped minimum version of ``tabulate`` to 0.9.*.
+
+**Fixed**
+
+- Fix ``:raises:`` field not supporting types.
+
+**Removed**
+
+- Removed support for Python 3.6.
+
 1.5.1 (2022/09/01)
 ------------------